Update logout logic

+++ src/app/components/pages/logout/logout.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { HttpServices } from 'src/app/components/connections/services/http-services';
 
 @Component({
   selector: 'app-logout',
@@ -8,12 +9,26 @@ import { Router } from '@angular/router';
 })
 export class LogoutComponent implements OnInit {
 
-  constructor(private router: Router) { }
+  constructor(
+  	private router: Router,
+    private _http: HttpServices,
+
+  ) { }
 
   ngOnInit(): void {
-    sessionStorage.clear()
-    console.log('Check---Loguot sesscesffuly>');
-    this.router.navigate(['/login'])
+  	this.logoutFromBackend()
+  }
+
+  logoutFromBackend(){
+    this._http.get('users/logout').subscribe((response: any) => {
+      if(response.status == 200){
+      	sessionStorage.clear()
+      	console.log('Check---Loguot sesscesffuly>');
+      	this.router.navigate(['/login'])
+      }
+    },err=>{
+      console.log(err)
+    })
   }
 
 }
